+++ packages/adherev-ui-searchtable/es/searchtableimplement.d.ts
@@ -1,3 +1,2 @@
-import Vue from 'vue';
-declare const _default: (serviceName: any) => import("vue/types/vue").ExtendedVue<Vue, any, any, unknown, Record<never, any>>;
+declare const _default: (serviceName: any) => any;
 export default _default;