+++ packages/adherev-util/lib/_util.d.ts
@@ -1,6 +1,15 @@
+import { ComponentOptions } from 'vue/types/options';
+/**
+ * ExtendFunction
+ */
+interface ExtendFunction<P> extends ComponentOptions<any> {
+    className?: string | string[];
+}
 /**
  * withInstall
+ * @description Vue组件的插件方法
  * @param comp
+ * @return comp
  */
 export declare const withInstall: (comp: any) => any;
 /**
@@ -12,8 +21,17 @@ export declare const withInstall: (comp: any) => any;
  */
 export declare const withVue: (Vue: any, p: string, val: any) => void;
 /**
+ * extend
+ * @description - 实现Vue的继承使得可以实现调用父类的方法
+ * 例如 类名A有一个renderInner方法，则父类的方法是$renderInnerA
+ * @param options
+ * @return Omit<ExtendFunction<any>, 'className'>
+ */
+export declare const extend: (options: ExtendFunction<any>) => Omit<ExtendFunction<any>, 'className'>;
+/**
  * Fragment
  */
 export declare const Fragment: {
     render(h: any): any;
 };
+export {};